Fix dlmalloc to compile without warnings on Mac OS.
Mac OS' GCC doesn't support #pragma GCC diagnostic.

+++ src/dlmalloc/malloc.c
@@ -4263,8 +4263,10 @@ static size_t release_unused_segments(mstate m) {
     sp = next;
   }
   /* Reset check counter */
-  m->release_checks = ((nsegs > MAX_RELEASE_CHECK_RATE)?
-                       nsegs : MAX_RELEASE_CHECK_RATE);
+  // BEGIN android-changed: signed/unsigned mismatches
+  m->release_checks = (((size_t) nsegs > (size_t) MAX_RELEASE_CHECK_RATE)?
+      (size_t) nsegs : (size_t) MAX_RELEASE_CHECK_RATE);
+  // END android-changed
   return released;
 }
